hashtbl.c: silence llvm-clang warning
llvm-clang warns about a tautological compare for enum type < 0. enum
types are not in general required to be unsigned, so just cast to int
before comparing to silence this silly warning.

+++ src/cwo/internal/hashtbl.c
@@ -59,7 +59,7 @@ cwoint_hashtbl_create(cwoint_hashtbl **self, enum cwoint_hashtbl_size size,
 {
     int err = CWO_SUCCESS;
 
-    if (size < 0 || size >= CWOINT_HASHTBL_NUMSIZES)
+    if ((int)size < 0 || size >= CWOINT_HASHTBL_NUMSIZES)
     {
 	return CWOIERR_HTSIZE;
     }
